+++ .storybook/main.ts
@@ -3,6 +3,7 @@ import type { StorybookConfig } from '@storybook/react-vite';
 const config: StorybookConfig = {
   stories: [
     '../docs/**/*.mdx',
+    '../docs/**/*.story.tsx',
     '../src/**/*.story.tsx'
   ],
   addons: [
