--- docs/blocks/administration/BillingInformation.tsx
@@ -1,311 +1,0 @@
-import React, { FC } from 'react';
-import { motion } from 'framer-motion';
-
-import { Card } from '../../../src/layout/Card';
-import { Divider } from '../../../src/layout/Divider';
-import { Button } from '../../../src/elements/Button';
-import {
-  Chip,
-  ChipTheme,
-  chipTheme as defaultChipTheme
-} from '../../../src/elements/Chip';
-
-export const BillingInformation: FC = () => {
-  const chipTheme: ChipTheme = {
-    ...defaultChipTheme,
-    base: `${defaultChipTheme?.base} rounded-full`,
-    colors: {
-      ...defaultChipTheme?.colors,
-      warning: {
-        ...defaultChipTheme.colors.warning,
-        variants: {
-          ...defaultChipTheme?.colors?.warning?.variants,
-          filled: `${defaultChipTheme?.colors?.warning?.variants?.filled} text-panel-content bg-warning/10 border-warning`
-        }
-      },
-      success: {
-        ...defaultChipTheme.colors.success,
-        variants: {
-          ...defaultChipTheme?.colors?.success?.variants,
-          filled: `${defaultChipTheme?.colors?.success?.variants?.filled} text-panel-content bg-success/10 border-success`
-        }
-      }
-    }
-  };
-
-  return (
-    <motion.div
-      initial={{ y: -20, opacity: 0 }}
-      animate={{ y: 0, opacity: 1 }}
-      className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm md:max-w-[1200px]"
-    >
-      <div className="flex justify-between gap-2 bg-primary/10 border border-primary rounded-xl p-8 text-panel-content mb-2">
-        <div className="flex flex-col gap-2">
-          <div className="flex flex-col">
-            <div className="flex gap-2 items-center">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="22"
-                height="22"
-                viewBox="0 0 22 22"
-                className="mb-0.5"
-                fill="currentColor"
-              >
-                <path d="M3.89583 4.58325C2.51177 4.58325 1.375 5.72002 1.375 7.10409V15.8124C1.375 17.1965 2.51177 18.3333 3.89583 18.3333H18.1042C19.4882 18.3333 20.625 17.1965 20.625 15.8124V7.10409C20.625 5.72002 19.4882 4.58325 18.1042 4.58325H3.89583ZM3.89583 5.95825H18.1042C18.745 5.95825 19.25 6.46323 19.25 7.10409V15.8124C19.25 16.4533 18.745 16.9583 18.1042 16.9583H3.89583C3.25498 16.9583 2.75 16.4533 2.75 15.8124V7.10409C2.75 6.46323 3.25498 5.95825 3.89583 5.95825ZM4.8125 7.79159C4.7214 7.7903 4.63095 7.80713 4.54641 7.8411C4.46187 7.87507 4.38492 7.92551 4.32004 7.98948C4.25516 8.05344 4.20364 8.12967 4.16848 8.21372C4.13332 8.29777 4.11521 8.38797 4.11521 8.47909C4.11521 8.5702 4.13332 8.6604 4.16848 8.74445C4.20364 8.8285 4.25516 8.90473 4.32004 8.9687C4.38492 9.03266 4.46187 9.0831 4.54641 9.11707C4.63095 9.15104 4.7214 9.16787 4.8125 9.16659H11.2292C11.3203 9.16787 11.4107 9.15104 11.4953 9.11707C11.5798 9.0831 11.6567 9.03266 11.7216 8.9687C11.7865 8.90473 11.838 8.8285 11.8732 8.74445C11.9084 8.6604 11.9265 8.5702 11.9265 8.47909C11.9265 8.38797 11.9084 8.29777 11.8732 8.21372C11.838 8.12967 11.7865 8.05344 11.7216 7.98948C11.6567 7.92551 11.5798 7.87507 11.4953 7.8411C11.4107 7.80713 11.3203 7.7903 11.2292 7.79159H4.8125ZM4.8125 10.5416C4.7214 10.5403 4.63095 10.5571 4.54641 10.5911C4.46187 10.6251 4.38492 10.6755 4.32004 10.7395C4.25516 10.8034 4.20364 10.8797 4.16848 10.9637C4.13332 11.0478 4.11521 11.138 4.11521 11.2291C4.11521 11.3202 4.13332 11.4104 4.16848 11.4945C4.20364 11.5785 4.25516 11.6547 4.32004 11.7187C4.38492 11.7827 4.46187 11.8331 4.54641 11.8671C4.63095 11.901 4.7214 11.9179 4.8125 11.9166H8.9375C9.0286 11.9179 9.11905 11.901 9.20359 11.8671C9.28813 11.8331 9.36508 11.7827 9.42996 11.7187C9.49484 11.6547 9.54636 11.5785 9.58152 11.4945C9.61668 11.4104 9.63479 11.3202 9.63479 11.2291C9.63479 11.138 9.61668 11.0478 9.58152 10.9637C9.54636 10.8797 9.49484 10.8034 9.42996 10.7395C9.36508 10.6755 9.28813 10.6251 9.20359 10.5911C9.11905 10.5571 9.0286 10.5403 8.9375 10.5416H4.8125ZM12.2604 11.9166C11.816 11.9166 11.529 12.1784 11.3402 12.4098C11.1514 12.6413 11.0196 12.8993 10.9069 13.1761C10.6815 13.7297 10.5417 14.3417 10.5417 14.8958C10.5404 14.9869 10.5572 15.0773 10.5912 15.1618C10.6252 15.2464 10.6756 15.3233 10.7396 15.3882C10.8035 15.4531 10.8797 15.5046 10.9638 15.5398C11.0479 15.5749 11.1381 15.593 11.2292 15.593C11.3203 15.593 11.4105 15.5749 11.4945 15.5398C11.5786 15.5046 11.6548 15.4531 11.7188 15.3882C11.7827 15.3233 11.8332 15.2464 11.8672 15.1618C11.9011 15.0773 11.918 14.9869 11.9167 14.8958C11.9167 14.612 12.024 14.0781 12.1799 13.6953C12.1924 13.6644 12.2034 13.6571 12.2166 13.6282C12.2676 13.7499 12.3131 13.8384 12.366 13.9845C12.4673 14.2638 12.5655 14.558 12.6937 14.8304C12.7578 14.9666 12.8252 15.0992 12.9452 15.2413C13.0653 15.3834 13.2917 15.5833 13.6354 15.5833C14.2885 15.5833 14.6509 15.1251 14.8573 14.8608C14.8599 14.8575 14.8601 14.8569 14.8627 14.8537C14.9189 14.9046 14.9335 14.9264 15.0095 14.9844C15.4191 15.2964 16.1057 15.5833 17.1875 15.5833C17.2786 15.5845 17.3691 15.5677 17.4536 15.5337C17.5381 15.4998 17.6151 15.4493 17.68 15.3854C17.7448 15.3214 17.7964 15.2452 17.8315 15.1611C17.8667 15.0771 17.8848 14.9869 17.8848 14.8958C17.8848 14.8046 17.8667 14.7144 17.8315 14.6304C17.7964 14.5463 17.7448 14.4701 17.68 14.4061C17.6151 14.3422 17.5381 14.2917 17.4536 14.2578C17.3691 14.2238 17.2786 14.207 17.1875 14.2083C16.3214 14.2083 16.034 14.0367 15.842 13.8905C15.746 13.8173 15.6739 13.7355 15.5386 13.613C15.4033 13.4904 15.125 13.2916 14.7812 13.2916C14.343 13.2916 14.1647 13.534 14.0472 13.6667C13.9516 13.7747 13.8891 13.8644 13.8252 13.9495C13.7709 13.8079 13.722 13.6901 13.6587 13.5154C13.5523 13.2218 13.4397 12.9118 13.2738 12.6211C13.1908 12.4757 13.0955 12.3312 12.9434 12.1932C12.7913 12.0551 12.5475 11.9166 12.2604 11.9166Z" />
-              </svg>
-              <span className="font-bold">Teams Plan</span>
-            </div>
-            <span className="text-base text-panel-secondary-content">
-              Organization plan - collaborate on projects with your team
-            </span>
-          </div>
-          <Button
-            className="w-fit font-semibold px-4 py-2 border-none bg-button-gradient hover:bg-button-gradient-hover focus:bg-button-gradient-focus light:bg-none light:bg-primary light:hover:bg-none light:hover:bg-primary-hover light:focus:bg-primary-hover focus:outline-none transition-colors"
-            color="primary"
-            endAdornment={
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="16"
-                height="16"
-                viewBox="0 0 16 16"
-                fill="currentColor"
-              >
-                <path d="M7.99999 2.66675V0.666748L5.33332 3.33341L7.99999 6.00008V4.00008C10.2067 4.00008 12 5.79342 12 8.00008C12 8.67342 11.8333 9.31342 11.5333 9.86675L12.5067 10.8401C13.0267 10.0201 13.3333 9.04675 13.3333 8.00008C13.3333 5.05341 10.9467 2.66675 7.99999 2.66675ZM7.99999 12.0001C5.79332 12.0001 3.99999 10.2067 3.99999 8.00008C3.99999 7.32675 4.16666 6.68675 4.46666 6.13341L3.49332 5.16008C2.97332 5.98008 2.66666 6.95341 2.66666 8.00008C2.66666 10.9467 5.05332 13.3334 7.99999 13.3334V15.3334L10.6667 12.6667L7.99999 10.0001V12.0001Z" />
-              </svg>
-            }
-          >
-            Change Plan
-          </Button>
-        </div>
-        <div className="flex flex-col">
-          <span className="font-bold">$62/mo</span>
-          <span className="text-base text-panel-secondary-content">
-            65 Seats
-          </span>
-        </div>
-      </div>
-      <Card className="transition-colors p-12">
-        <div className="flex flex-col gap-2">
-          <h2 className="font-bold text-2xl">Billing Information</h2>
-          <Divider variant="secondary" className="mb-4" />
-          <div className="flex w-full gap-4">
-            <Card contentClassName="flex flex-col gap-4" className="flex-1">
-              <div className="flex justify-between items-center">
-                <span className="font-bold">Payment Method</span>
-                <Button
-                  className="w-fit font-semibold px-4 py-2 border-none dark:bg-charade/40 dark:hover:bg-charade/60 dark:focus:bg-charade/60 light:bg-blue-200/40 light:hover:bg-blue-200/60 light:focus:bg-blue-200/60 light:text-vulcan focus:outline-none transition-colors"
-                  color="secondary"
-                >
-                  Update
-                </Button>
-              </div>
-              <div className="flex flex-col">
-                <span className="text-sm text-panel-secondary-content">
-                  Card information
-                </span>
-                <span className="font-bold">Visa ending in 0315</span>
-              </div>
-              <div className="flex flex-col">
-                <span className="text-sm text-panel-secondary-content">
-                  Name on card
-                </span>
-                <span className="font-bold">Austin McDaniel</span>
-              </div>
-            </Card>
-            <Card contentClassName="flex flex-col gap-4" className="flex-1">
-              <div className="flex justify-between items-center">
-                <span className="font-bold">Billing Details</span>
-                <Button
-                  className="w-fit font-semibold px-4 py-2 border-none dark:bg-charade/40 dark:hover:bg-charade/60 dark:focus:bg-charade/60 light:bg-blue-200/40 light:hover:bg-blue-200/60 light:focus:bg-blue-200/60 light:text-vulcan focus:outline-none transition-colors"
-                  color="secondary"
-                >
-                  Update
-                </Button>
-              </div>
-              <div className="flex flex-col">
-                <span className="text-sm text-panel-secondary-content">
-                  Next billing date
-                </span>
-                <span className="font-bold">Mar 15, 2025</span>
-              </div>
-              <div className="flex flex-col">
-                <span className="text-sm text-panel-secondary-content">
-                  Billing address
-                </span>
-                <span className="font-bold">
-                  123 Florida Ave., Miami FL, 33125
-                </span>
-              </div>
-            </Card>
-          </div>
-          <div className="flex w-full">
-            <div className="flex flex-col flex-1">
-              <div className="border-b border-panel-accent py-2 px-4">
-                <span className="text-panel-secondary-content text-base">
-                  Invoice Date
-                </span>
-              </div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.1 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <span className="font-bold">Mar 15, 2023</span>
-              </motion.div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.2 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <span className="font-bold">Mar 15, 2022</span>
-              </motion.div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.3 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <span className="font-bold">Mar 15, 2021</span>
-              </motion.div>
-            </div>
-            <div className="flex flex-col flex-1">
-              <div className="border-b border-panel-accent py-2 px-4">
-                <span className="text-panel-secondary-content text-base">
-                  Status
-                </span>
-              </div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.1 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <Chip theme={chipTheme} color="warning" variant="filled">
-                  Pending
-                </Chip>
-              </motion.div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.2 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <Chip theme={chipTheme} color="success" variant="filled">
-                  Paid
-                </Chip>
-              </motion.div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.3 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <Chip theme={chipTheme} color="success" variant="filled">
-                  Paid
-                </Chip>
-              </motion.div>
-            </div>
-            <div className="flex flex-col flex-1">
-              <div className="border-b border-panel-accent py-2 px-4">
-                <span className="text-panel-secondary-content text-base">
-                  Card Used
-                </span>
-              </div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.1 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <div className="flex gap-2 items-center">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="16"
-                    height="16"
-                    viewBox="0 0 16 16"
-                    fill="currentColor"
-                  >
-                    <path d="M2.83333 3.33325C1.82674 3.33325 1 4.15999 1 5.16659V11.4999C1 12.5065 1.82674 13.3333 2.83333 13.3333H13.1667C14.1733 13.3333 15 12.5065 15 11.4999V5.16659C15 4.15999 14.1733 3.33325 13.1667 3.33325H2.83333ZM2.83333 4.33325H13.1667C13.6327 4.33325 14 4.70051 14 5.16659V11.4999C14 11.966 13.6327 12.3333 13.1667 12.3333H2.83333C2.36726 12.3333 2 11.966 2 11.4999V5.16659C2 4.70051 2.36726 4.33325 2.83333 4.33325ZM3.5 5.66659C3.43374 5.66565 3.36796 5.67789 3.30648 5.7026C3.24499 5.7273 3.18903 5.76398 3.14185 5.81051C3.09466 5.85703 3.0572 5.91246 3.03162 5.97359C3.00605 6.03472 2.99288 6.10032 2.99288 6.16659C2.99288 6.23285 3.00605 6.29845 3.03162 6.35958C3.0572 6.42071 3.09466 6.47614 3.14185 6.52267C3.18903 6.56919 3.24499 6.60587 3.30648 6.63057C3.36796 6.65528 3.43374 6.66752 3.5 6.66659H8.16667C8.23292 6.66752 8.2987 6.65528 8.36019 6.63057C8.42167 6.60587 8.47763 6.56919 8.52482 6.52267C8.572 6.47614 8.60947 6.42071 8.63505 6.35958C8.66062 6.29845 8.67379 6.23285 8.67379 6.16659C8.67379 6.10032 8.66062 6.03472 8.63505 5.97359C8.60947 5.91246 8.572 5.85703 8.52482 5.81051C8.47763 5.76398 8.42167 5.7273 8.36019 5.7026C8.2987 5.67789 8.23292 5.66565 8.16667 5.66659H3.5ZM3.5 7.66659C3.43374 7.66565 3.36796 7.67789 3.30648 7.7026C3.24499 7.7273 3.18903 7.76398 3.14185 7.81051C3.09466 7.85703 3.0572 7.91246 3.03162 7.97359C3.00605 8.03472 2.99288 8.10032 2.99288 8.16659C2.99288 8.23285 3.00605 8.29845 3.03162 8.35958C3.0572 8.42071 3.09466 8.47614 3.14185 8.52267C3.18903 8.56919 3.24499 8.60587 3.30648 8.63057C3.36796 8.65528 3.43374 8.66752 3.5 8.66659H6.5C6.56626 8.66752 6.63204 8.65528 6.69352 8.63057C6.75501 8.60587 6.81097 8.56919 6.85815 8.52267C6.90534 8.47614 6.9428 8.42071 6.96838 8.35958C6.99395 8.29845 7.00712 8.23285 7.00712 8.16659C7.00712 8.10032 6.99395 8.03472 6.96838 7.97359C6.9428 7.91246 6.90534 7.85703 6.85815 7.81051C6.81097 7.76398 6.75501 7.7273 6.69352 7.7026C6.63204 7.67789 6.56626 7.66565 6.5 7.66659H3.5ZM8.91667 8.66659C8.59346 8.66659 8.3847 8.85699 8.2474 9.02531C8.11009 9.19363 8.01427 9.3813 7.93229 9.5826C7.76834 9.9852 7.66667 10.4303 7.66667 10.8333C7.66573 10.8995 7.67797 10.9653 7.70268 11.0268C7.72738 11.0883 7.76406 11.1442 7.81059 11.1914C7.85711 11.2386 7.91255 11.2761 7.97367 11.3016C8.0348 11.3272 8.1004 11.3404 8.16667 11.3404C8.23293 11.3404 8.29853 11.3272 8.35966 11.3016C8.42079 11.2761 8.47622 11.2386 8.52275 11.1914C8.56927 11.1442 8.60595 11.0883 8.63066 11.0268C8.65536 10.9653 8.6676 10.8995 8.66667 10.8333C8.66667 10.6269 8.7447 10.2386 8.85807 9.96021C8.86723 9.93771 8.87521 9.93241 8.88477 9.91138C8.92186 9.99989 8.95499 10.0643 8.99349 10.1705C9.06713 10.3736 9.13851 10.5876 9.23177 10.7857C9.2784 10.8848 9.32742 10.9812 9.41471 11.0846C9.50201 11.1879 9.66667 11.3333 9.91667 11.3333C10.3917 11.3333 10.6552 11 10.8053 10.8079C10.8072 10.8055 10.8074 10.805 10.8092 10.8027C10.8501 10.8397 10.8607 10.8556 10.916 10.8977C11.2139 11.1247 11.7132 11.3333 12.5 11.3333C12.5663 11.3342 12.632 11.3219 12.6935 11.2972C12.755 11.2725 12.811 11.2359 12.8582 11.1893C12.9053 11.1428 12.9428 11.0874 12.9684 11.0262C12.994 10.9651 13.0071 10.8995 13.0071 10.8333C13.0071 10.767 12.994 10.7014 12.9684 10.6403C12.9428 10.5791 12.9053 10.5237 12.8582 10.4772C12.811 10.4307 12.755 10.394 12.6935 10.3693C12.632 10.3446 12.5663 10.3323 12.5 10.3333C11.8701 10.3333 11.6611 10.2085 11.5215 10.1021C11.4517 10.0489 11.3992 9.98942 11.3008 9.90031C11.2024 9.8112 11 9.66659 10.75 9.66659C10.4313 9.66659 10.3016 9.84287 10.2161 9.93937C10.1466 10.0179 10.1012 10.0832 10.0547 10.1451C10.0152 10.0421 9.97967 9.95644 9.93359 9.82935C9.85619 9.61582 9.77434 9.3904 9.65365 9.17896C9.5933 9.07323 9.52403 8.96816 9.41341 8.86776C9.30279 8.76736 9.12547 8.66659 8.91667 8.66659Z" />
-                  </svg>
-                  <span className="text-panel-secondary-content text-base">
-                    Visa ending in 0315
-                  </span>
-                </div>
-              </motion.div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.2 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <div className="flex gap-2 justify-between items-center">
-                  <div className="flex gap-2 items-center">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="16"
-                      height="16"
-                      viewBox="0 0 16 16"
-                      fill="currentColor"
-                    >
-                      <path d="M2.83333 3.33325C1.82674 3.33325 1 4.15999 1 5.16659V11.4999C1 12.5065 1.82674 13.3333 2.83333 13.3333H13.1667C14.1733 13.3333 15 12.5065 15 11.4999V5.16659C15 4.15999 14.1733 3.33325 13.1667 3.33325H2.83333ZM2.83333 4.33325H13.1667C13.6327 4.33325 14 4.70051 14 5.16659V11.4999C14 11.966 13.6327 12.3333 13.1667 12.3333H2.83333C2.36726 12.3333 2 11.966 2 11.4999V5.16659C2 4.70051 2.36726 4.33325 2.83333 4.33325ZM3.5 5.66659C3.43374 5.66565 3.36796 5.67789 3.30648 5.7026C3.24499 5.7273 3.18903 5.76398 3.14185 5.81051C3.09466 5.85703 3.0572 5.91246 3.03162 5.97359C3.00605 6.03472 2.99288 6.10032 2.99288 6.16659C2.99288 6.23285 3.00605 6.29845 3.03162 6.35958C3.0572 6.42071 3.09466 6.47614 3.14185 6.52267C3.18903 6.56919 3.24499 6.60587 3.30648 6.63057C3.36796 6.65528 3.43374 6.66752 3.5 6.66659H8.16667C8.23292 6.66752 8.2987 6.65528 8.36019 6.63057C8.42167 6.60587 8.47763 6.56919 8.52482 6.52267C8.572 6.47614 8.60947 6.42071 8.63505 6.35958C8.66062 6.29845 8.67379 6.23285 8.67379 6.16659C8.67379 6.10032 8.66062 6.03472 8.63505 5.97359C8.60947 5.91246 8.572 5.85703 8.52482 5.81051C8.47763 5.76398 8.42167 5.7273 8.36019 5.7026C8.2987 5.67789 8.23292 5.66565 8.16667 5.66659H3.5ZM3.5 7.66659C3.43374 7.66565 3.36796 7.67789 3.30648 7.7026C3.24499 7.7273 3.18903 7.76398 3.14185 7.81051C3.09466 7.85703 3.0572 7.91246 3.03162 7.97359C3.00605 8.03472 2.99288 8.10032 2.99288 8.16659C2.99288 8.23285 3.00605 8.29845 3.03162 8.35958C3.0572 8.42071 3.09466 8.47614 3.14185 8.52267C3.18903 8.56919 3.24499 8.60587 3.30648 8.63057C3.36796 8.65528 3.43374 8.66752 3.5 8.66659H6.5C6.56626 8.66752 6.63204 8.65528 6.69352 8.63057C6.75501 8.60587 6.81097 8.56919 6.85815 8.52267C6.90534 8.47614 6.9428 8.42071 6.96838 8.35958C6.99395 8.29845 7.00712 8.23285 7.00712 8.16659C7.00712 8.10032 6.99395 8.03472 6.96838 7.97359C6.9428 7.91246 6.90534 7.85703 6.85815 7.81051C6.81097 7.76398 6.75501 7.7273 6.69352 7.7026C6.63204 7.67789 6.56626 7.66565 6.5 7.66659H3.5ZM8.91667 8.66659C8.59346 8.66659 8.3847 8.85699 8.2474 9.02531C8.11009 9.19363 8.01427 9.3813 7.93229 9.5826C7.76834 9.9852 7.66667 10.4303 7.66667 10.8333C7.66573 10.8995 7.67797 10.9653 7.70268 11.0268C7.72738 11.0883 7.76406 11.1442 7.81059 11.1914C7.85711 11.2386 7.91255 11.2761 7.97367 11.3016C8.0348 11.3272 8.1004 11.3404 8.16667 11.3404C8.23293 11.3404 8.29853 11.3272 8.35966 11.3016C8.42079 11.2761 8.47622 11.2386 8.52275 11.1914C8.56927 11.1442 8.60595 11.0883 8.63066 11.0268C8.65536 10.9653 8.6676 10.8995 8.66667 10.8333C8.66667 10.6269 8.7447 10.2386 8.85807 9.96021C8.86723 9.93771 8.87521 9.93241 8.88477 9.91138C8.92186 9.99989 8.95499 10.0643 8.99349 10.1705C9.06713 10.3736 9.13851 10.5876 9.23177 10.7857C9.2784 10.8848 9.32742 10.9812 9.41471 11.0846C9.50201 11.1879 9.66667 11.3333 9.91667 11.3333C10.3917 11.3333 10.6552 11 10.8053 10.8079C10.8072 10.8055 10.8074 10.805 10.8092 10.8027C10.8501 10.8397 10.8607 10.8556 10.916 10.8977C11.2139 11.1247 11.7132 11.3333 12.5 11.3333C12.5663 11.3342 12.632 11.3219 12.6935 11.2972C12.755 11.2725 12.811 11.2359 12.8582 11.1893C12.9053 11.1428 12.9428 11.0874 12.9684 11.0262C12.994 10.9651 13.0071 10.8995 13.0071 10.8333C13.0071 10.767 12.994 10.7014 12.9684 10.6403C12.9428 10.5791 12.9053 10.5237 12.8582 10.4772C12.811 10.4307 12.755 10.394 12.6935 10.3693C12.632 10.3446 12.5663 10.3323 12.5 10.3333C11.8701 10.3333 11.6611 10.2085 11.5215 10.1021C11.4517 10.0489 11.3992 9.98942 11.3008 9.90031C11.2024 9.8112 11 9.66659 10.75 9.66659C10.4313 9.66659 10.3016 9.84287 10.2161 9.93937C10.1466 10.0179 10.1012 10.0832 10.0547 10.1451C10.0152 10.0421 9.97967 9.95644 9.93359 9.82935C9.85619 9.61582 9.77434 9.3904 9.65365 9.17896C9.5933 9.07323 9.52403 8.96816 9.41341 8.86776C9.30279 8.76736 9.12547 8.66659 8.91667 8.66659Z" />
-                    </svg>
-                    <span className="text-panel-secondary-content text-base">
-                      Visa ending in 0315
-                    </span>
-                  </div>
-                  <Button variant="text" className="p-0.5">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="16"
-                      height="16"
-                      viewBox="0 0 16 16"
-                      fill="currentColor"
-                    >
-                      <path d="M12.6666 6.33325H9.99992V2.33325H5.99992V6.33325H3.33325L7.99992 10.9999L12.6666 6.33325ZM7.33325 7.66658V3.66659H8.66658V7.66658H9.44659L7.99992 9.11325L6.55325 7.66658H7.33325ZM3.33325 12.3333H12.6666V13.6666H3.33325V12.3333Z" />
-                    </svg>
-                  </Button>
-                </div>
-              </motion.div>
-              <motion.div
-                initial={{ y: -20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ delay: 0.3 }}
-                className="border-b border-panel-accent py-5 px-4 flex-1"
-              >
-                <div className="flex gap-2 justify-between items-center">
-                  <div className="flex gap-2 items-center">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="16"
-                      height="16"
-                      viewBox="0 0 16 16"
-                      fill="currentColor"
-                    >
-                      <path d="M2.83333 3.33325C1.82674 3.33325 1 4.15999 1 5.16659V11.4999C1 12.5065 1.82674 13.3333 2.83333 13.3333H13.1667C14.1733 13.3333 15 12.5065 15 11.4999V5.16659C15 4.15999 14.1733 3.33325 13.1667 3.33325H2.83333ZM2.83333 4.33325H13.1667C13.6327 4.33325 14 4.70051 14 5.16659V11.4999C14 11.966 13.6327 12.3333 13.1667 12.3333H2.83333C2.36726 12.3333 2 11.966 2 11.4999V5.16659C2 4.70051 2.36726 4.33325 2.83333 4.33325ZM3.5 5.66659C3.43374 5.66565 3.36796 5.67789 3.30648 5.7026C3.24499 5.7273 3.18903 5.76398 3.14185 5.81051C3.09466 5.85703 3.0572 5.91246 3.03162 5.97359C3.00605 6.03472 2.99288 6.10032 2.99288 6.16659C2.99288 6.23285 3.00605 6.29845 3.03162 6.35958C3.0572 6.42071 3.09466 6.47614 3.14185 6.52267C3.18903 6.56919 3.24499 6.60587 3.30648 6.63057C3.36796 6.65528 3.43374 6.66752 3.5 6.66659H8.16667C8.23292 6.66752 8.2987 6.65528 8.36019 6.63057C8.42167 6.60587 8.47763 6.56919 8.52482 6.52267C8.572 6.47614 8.60947 6.42071 8.63505 6.35958C8.66062 6.29845 8.67379 6.23285 8.67379 6.16659C8.67379 6.10032 8.66062 6.03472 8.63505 5.97359C8.60947 5.91246 8.572 5.85703 8.52482 5.81051C8.47763 5.76398 8.42167 5.7273 8.36019 5.7026C8.2987 5.67789 8.23292 5.66565 8.16667 5.66659H3.5ZM3.5 7.66659C3.43374 7.66565 3.36796 7.67789 3.30648 7.7026C3.24499 7.7273 3.18903 7.76398 3.14185 7.81051C3.09466 7.85703 3.0572 7.91246 3.03162 7.97359C3.00605 8.03472 2.99288 8.10032 2.99288 8.16659C2.99288 8.23285 3.00605 8.29845 3.03162 8.35958C3.0572 8.42071 3.09466 8.47614 3.14185 8.52267C3.18903 8.56919 3.24499 8.60587 3.30648 8.63057C3.36796 8.65528 3.43374 8.66752 3.5 8.66659H6.5C6.56626 8.66752 6.63204 8.65528 6.69352 8.63057C6.75501 8.60587 6.81097 8.56919 6.85815 8.52267C6.90534 8.47614 6.9428 8.42071 6.96838 8.35958C6.99395 8.29845 7.00712 8.23285 7.00712 8.16659C7.00712 8.10032 6.99395 8.03472 6.96838 7.97359C6.9428 7.91246 6.90534 7.85703 6.85815 7.81051C6.81097 7.76398 6.75501 7.7273 6.69352 7.7026C6.63204 7.67789 6.56626 7.66565 6.5 7.66659H3.5ZM8.91667 8.66659C8.59346 8.66659 8.3847 8.85699 8.2474 9.02531C8.11009 9.19363 8.01427 9.3813 7.93229 9.5826C7.76834 9.9852 7.66667 10.4303 7.66667 10.8333C7.66573 10.8995 7.67797 10.9653 7.70268 11.0268C7.72738 11.0883 7.76406 11.1442 7.81059 11.1914C7.85711 11.2386 7.91255 11.2761 7.97367 11.3016C8.0348 11.3272 8.1004 11.3404 8.16667 11.3404C8.23293 11.3404 8.29853 11.3272 8.35966 11.3016C8.42079 11.2761 8.47622 11.2386 8.52275 11.1914C8.56927 11.1442 8.60595 11.0883 8.63066 11.0268C8.65536 10.9653 8.6676 10.8995 8.66667 10.8333C8.66667 10.6269 8.7447 10.2386 8.85807 9.96021C8.86723 9.93771 8.87521 9.93241 8.88477 9.91138C8.92186 9.99989 8.95499 10.0643 8.99349 10.1705C9.06713 10.3736 9.13851 10.5876 9.23177 10.7857C9.2784 10.8848 9.32742 10.9812 9.41471 11.0846C9.50201 11.1879 9.66667 11.3333 9.91667 11.3333C10.3917 11.3333 10.6552 11 10.8053 10.8079C10.8072 10.8055 10.8074 10.805 10.8092 10.8027C10.8501 10.8397 10.8607 10.8556 10.916 10.8977C11.2139 11.1247 11.7132 11.3333 12.5 11.3333C12.5663 11.3342 12.632 11.3219 12.6935 11.2972C12.755 11.2725 12.811 11.2359 12.8582 11.1893C12.9053 11.1428 12.9428 11.0874 12.9684 11.0262C12.994 10.9651 13.0071 10.8995 13.0071 10.8333C13.0071 10.767 12.994 10.7014 12.9684 10.6403C12.9428 10.5791 12.9053 10.5237 12.8582 10.4772C12.811 10.4307 12.755 10.394 12.6935 10.3693C12.632 10.3446 12.5663 10.3323 12.5 10.3333C11.8701 10.3333 11.6611 10.2085 11.5215 10.1021C11.4517 10.0489 11.3992 9.98942 11.3008 9.90031C11.2024 9.8112 11 9.66659 10.75 9.66659C10.4313 9.66659 10.3016 9.84287 10.2161 9.93937C10.1466 10.0179 10.1012 10.0832 10.0547 10.1451C10.0152 10.0421 9.97967 9.95644 9.93359 9.82935C9.85619 9.61582 9.77434 9.3904 9.65365 9.17896C9.5933 9.07323 9.52403 8.96816 9.41341 8.86776C9.30279 8.76736 9.12547 8.66659 8.91667 8.66659Z" />
-                    </svg>
-                    <span className="text-panel-secondary-content text-base">
-                      Visa ending in 0315
-                    </span>
-                  </div>
-                  <Button variant="text" className="p-0.5">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="16"
-                      height="16"
-                      viewBox="0 0 16 16"
-                      fill="currentColor"
-                    >
-                      <path d="M12.6666 6.33325H9.99992V2.33325H5.99992V6.33325H3.33325L7.99992 10.9999L12.6666 6.33325ZM7.33325 7.66658V3.66659H8.66658V7.66658H9.44659L7.99992 9.11325L6.55325 7.66658H7.33325ZM3.33325 12.3333H12.6666V13.6666H3.33325V12.3333Z" />
-                    </svg>
-                  </Button>
-                </div>
-              </motion.div>
-            </div>
-          </div>
-        </div>
-      </Card>
-    </motion.div>
-  );
-};
